requirements.txt
sentence_transformers==3.4.1
faiss-cpu==1.10.0
numpy==1.26.0
pandas==2.2.3
pyodbc==5.2.0
ipykernel==6.29.4
requests==2.31.0
ipywidgets==8.1.5
requests==2.31.0
scikit-learn==1.3.1
streamlit==1.38.0
matplotlib_scalebar==0.8.1
plotly==5.22.0
python-dotenv==1.0.0
